test: add html template test
Added comments with painfully calculated byte indexes. Start is at the open brackets, or right after. End is at the close brackets, or right before the open bracket of the end comment. Included Windows calculations, but they did not appear relevant when previously implementing this, so I do not think they'll be relevant now. Including in case additional changes will need to be made. Everything appears to work correctly now.